# docs/swagger.yaml
definitions:
...
  internal.DetailPermission:
    properties:
      action:
        type: string
      content_ids:
        items:
          type: integer
        type: array
      content_type:
        type: string
      id:
        type: integer
      user_id:
        type: integer
    type: object
  internal.Permission:
    properties:
      action:
        type: string
      content_type:
        type: string
      description:
        type: string
      detail:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/detail_permissions:
    get:
      consumes:
      - application/json
      description: Retrieve list of detail permissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal.DetailPermission'
            type: array
      tags:
      - permissions
    post:
      consumes:
      - application/json
      description: Create a new detail permission
      parameters:
      - description: Detail Permission
        in: body
        name: detailPermission
        required: true
        schema:
          $ref: '#/definitions/internal.DetailPermission'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal.DetailPermission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - permissions
  /api/detail_permissions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a detail permission by ID
      parameters:
      - description: Detail Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - permissions
    get:
      consumes:
      - application/json
      description: Retrieve a detail permission by ID
      parameters:
      - description: Detail Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.DetailPermission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - permissions
    put:
      consumes:
      - application/json
      description: Update an existing detail permission
      parameters:
      - description: Detail Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Detail Permission
        in: body
        name: detailPermission
        required: true
        schema:
          $ref: '#/definitions/internal.DetailPermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.DetailPermission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - permissions
  /api/permissions:
    get:
      consumes:
      - application/json
      description: Retrieve list of permissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal.Permission'
            type: array
      tags:
      - permissions
    post:
      consumes:
      - application/json
      description: Create a new permission
      parameters:
      - description: Permission
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/internal.Permission'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal.Permission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - permissions
  /api/permissions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a permission by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - permissions
    get:
      consumes:
      - application/json
      description: Retrieve a permission by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Permission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - permissions
    put:
      consumes:
      - application/json
      description: Update an existing permission
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/internal.Permission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Permission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - permissions
...
swagger: "2.0"